Refactor: move around reaperd code
The `reaper` module had both node and pod stuff, which were only called
from the respecitve `*_watcher` modules.  Better to just have more
self-contained `pod_reaper` and `node_reaper` modules.
