+++ crates/reaperd/src/main.rs
@@ -4,8 +4,8 @@ use std::process;
 use nodelib::etcd;
 use nodelib::resources::node::*;
 
-use reaperd::node_watcher;
-use reaperd::pod_watcher;
+use reaperd::node_reaper;
+use reaperd::pod_reaper;
 
 /// thing-doer reaperd.
 #[derive(Clone, Debug, Parser)]
@@ -34,8 +34,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     )
     .await?;
 
-    let reap_pod_tx = pod_watcher::initialise(etcd.clone(), state.name.clone()).await?;
-    node_watcher::initialise(etcd, reap_pod_tx).await?;
+    let reap_pod_tx = pod_reaper::initialise(etcd.clone(), state.name.clone()).await?;
+    node_reaper::initialise(etcd, reap_pod_tx).await?;
 
     let ch = nodelib::wait_for_sigterm(state).await;
     nodelib::signal_channel(ch).await;
