+++ crates/reaperd/src/node_reaper.rs
@@ -2,15 +2,16 @@ use std::sync::Arc;
 use std::time::Duration;
 use tokio::sync::mpsc;
 use tokio::sync::RwLock;
+use tonic::Request;
 
 use nodelib::error::*;
 use nodelib::etcd;
+use nodelib::etcd::pb::etcdserverpb::range_request::{SortOrder, SortTarget};
+use nodelib::etcd::pb::etcdserverpb::{DeleteRangeRequest, RangeRequest};
 use nodelib::etcd::pb::mvccpb::{event::EventType, Event};
 use nodelib::etcd::prefix;
 use nodelib::etcd::watcher;
 use nodelib::resources::node::NodeType;
-
-use crate::reaper::drain_node_inbox;
 
 /// Interval to retry pod that could not be reaped.
 pub static RETRY_INTERVAL: u64 = 60;
@@ -123,4 +124,84 @@ async fn enqueue_retries(to_retry: Vec<String>, reap_node_tx: &mpsc::UnboundedSe
             .send(node_name)
             .expect("could not send to unbounded channel");
     }
+}
+
+///////////////////////////////////////////////////////////////////////////////
+
+/// Enqueue all pods in a node's inbox to be marked as dead.
+pub async fn drain_node_inbox(
+    etcd_config: &etcd::Config,
+    reap_pod_tx: mpsc::UnboundedSender<String>,
+    node_name: &str,
+) -> Result<bool, Error> {
+    let to_reap = get_inbox_for_node(etcd_config, node_name).await?;
+
+    if to_reap.is_empty() {
+        return Ok(false);
+    }
+
+    let mut kv_client = etcd_config.kv_client().await?;
+    for pod_name in to_reap {
+        tracing::info!(node_name, pod_name, "found reapable pod");
+        reap_pod_tx
+            .send(pod_name.to_owned())
+            .expect("could not send to unbounded channel");
+        kv_client
+            .delete_range(Request::new(DeleteRangeRequest {
+                key: format!(
+                    "{prefix}{pod_name}",
+                    prefix = prefix::worker_inbox(etcd_config, node_name)
+                )
+                .into(),
+                ..Default::default()
+            }))
+            .await?;
+    }
+
+    Ok(true)
+}
+
+/// Get all pods in a worker's inbox (i.e. pods that have been scheduled but not
+/// picked up).
+async fn get_inbox_for_node(
+    etcd_config: &etcd::Config,
+    node_name: &str,
+) -> Result<Vec<String>, Error> {
+    let mut kv_client = etcd_config.kv_client().await?;
+
+    let key_prefix = prefix::worker_inbox(etcd_config, node_name);
+    let range_end = prefix::range_end(&key_prefix);
+
+    let mut to_reap = Vec::with_capacity(128);
+    let mut key = key_prefix.as_bytes().to_vec();
+    let mut revision = 0;
+    loop {
+        let response = kv_client
+            .range(Request::new(RangeRequest {
+                key,
+                range_end: range_end.clone(),
+                revision,
+                sort_order: SortOrder::Ascend.into(),
+                sort_target: SortTarget::Key.into(),
+                ..Default::default()
+            }))
+            .await?
+            .into_inner();
+        revision = response.header.unwrap().revision;
+
+        for kv in &response.kvs {
+            let pod_key = String::from_utf8(kv.key.clone()).unwrap();
+            let (_, pod_name) = pod_key.split_once(&key_prefix).unwrap();
+            to_reap.push(pod_name.to_owned());
+        }
+
+        if response.more {
+            let idx = response.kvs.len() - 1;
+            key = response.kvs[idx].key.clone();
+        } else {
+            break;
+        }
+    }
+
+    Ok(to_reap)
 }
